making body of api gateway input event generic
The body is actually a string, but could be any value that serializes to
a string. I think the most common use is probably a JSON value, so
making body generic, but constrained to `FromJSON`. Implementors would
need to write a FromJSON instance for whatever value they expect or
simply use `APIGatewayInputEvent Text`.